IMPCAP::Fixes: segfault, memory and build corrections (resolves #4332)
- fix bug in ethernet packets parsing
  - fix removes build error with gcc10: 'multiple definition of...'
- resolve memory leak during interface init failure (device not freed after post-create error)
- add test 'impcap_bug_ether' to prove ethernet parser fix is working

+++ contrib/impcap/dns_parser.c
@@ -316,7 +316,7 @@ data_ret_t *dns_parse(const uchar *packet, int pktSize, struct json_object *jpar
 		char domain_query[256] = {0};
 		uchar nb_char = *packet_ptr;
 		packet_ptr++;
-		int cpt = 0;
+		size_t cpt = 0;
 		while (cpt + 1 < query_size) {
 			if (nb_char == 0) {
 				nb_char = *packet_ptr;
